IGNITE-17365 Move internal indexing classes related to H2 to another package
