+++ modules/indexing/src/main/java/org/apache/ignite/internal/processors/query/h2/mxbean/SqlQueryMXBean.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.ignite.internal.mxbean;
+package org.apache.ignite.internal.processors.query.h2.mxbean;
 
 import org.apache.ignite.mxbean.MXBeanDescription;
 import org.apache.ignite.mxbean.MXBeanParameter;
